backend/requirements.txt
flask
requests
geocoder
kafka-python
tweepy==3.7.0
folium
pandas
textblob
geopy